Add Username Validation to Server

+++ routes/ideas.js
@@ -44,17 +44,23 @@ router.post('/', async (req, res) => {
 // put/edit idea
 router.put('/:id', async (req, res) => {
 	try {
-		const UpdatedIdea = await Idea.findByIdAndUpdate(
-			req.params.id,
-			{
-				$set: {
-					text: req.body.text,
-					tag: req.body.tag,
+		const idea = await Idea.findById(req.params.id);
+
+		if (idea.username === req.body.username) {
+			const UpdatedIdea = await Idea.findByIdAndUpdate(
+				req.params.id,
+				{
+					$set: {
+						text: req.body.text,
+						tag: req.body.tag,
+					},
 				},
-			},
-			{ new: true }
-		);
-		res.json({ success: true, data: UpdatedIdea });
+				{ new: true }
+			);
+			return res.json({ success: true, data: UpdatedIdea });
+		}
+		// username does not match
+		res.status(403).json({ success: false, error: error.message });
 	} catch (error) {
 		console.error('Error in updating idea:', error); // log full error
 		res.status(500).json({ success: false, error: error.message });
@@ -64,8 +70,15 @@ router.put('/:id', async (req, res) => {
 // delete an idea
 router.delete('/:id', async (req, res) => {
 	try {
-		await Idea.findByIdAndDelete(req.params.id);
-		res.json({ success: true, data: {} });
+		const idea = await Idea.findById(req.params.id);
+
+		// match id username
+		if (idea.username === req.body.username) {
+			await Idea.findByIdAndDelete(req.params.id);
+			res.json({ success: true, data: {} });
+		} else {
+			res.status(403).json({ success: false, error: error.message });
+		}
 	} catch (error) {
 		console.error('Error in updating idea:', error); // log full error
 		res.status(500).json({ success: false, error: error.message });
